refactor: extract BicycleRentalStationRowContainer
